Title: 10개 입력받기, Time: 135ms, Memory: 10MB, Status: Passed - CodeTree

+++ 241031/10개 입력받기/receive-10-inputs.java
@@ -17,7 +17,7 @@ public class Main {
 
             sum+=arr[i];
             int cnt = i+1;
-            average=sum/cnt;
+            average=(double)sum/cnt;
 
         }
